+++ assetconnection/common/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/assetconnection/common/format/JsonFormat.java
@@ -14,10 +14,7 @@
  */
 package de.fraunhofer.iosb.ilt.faaast.service.assetconnection.common.format;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.io.JsonStringEncoder;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jayway.jsonpath.Configuration;
 import com.jayway.jsonpath.InvalidPathException;
 import com.jayway.jsonpath.JsonPath;
@@ -30,22 +27,21 @@ import de.fraunhofer.iosb.ilt.faaast.service.dataformat.DeserializationException
 import de.fraunhofer.iosb.ilt.faaast.service.dataformat.SerializationException;
 import de.fraunhofer.iosb.ilt.faaast.service.dataformat.json.JsonApiDeserializer;
 import de.fraunhofer.iosb.ilt.faaast.service.dataformat.json.JsonApiSerializer;
+import de.fraunhofer.iosb.ilt.faaast.service.model.api.modifier.Content;
+import de.fraunhofer.iosb.ilt.faaast.service.model.api.modifier.Extent;
+import de.fraunhofer.iosb.ilt.faaast.service.model.api.modifier.Level;
+import de.fraunhofer.iosb.ilt.faaast.service.model.api.modifier.OutputModifier;
 import de.fraunhofer.iosb.ilt.faaast.service.model.exception.UnsupportedModifierException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.value.DataElementValue;
 import de.fraunhofer.iosb.ilt.faaast.service.model.value.Datatype;
 import de.fraunhofer.iosb.ilt.faaast.service.typing.ElementValueTypeInfo;
 import de.fraunhofer.iosb.ilt.faaast.service.typing.TypeInfo;
 import de.fraunhofer.iosb.ilt.faaast.service.util.LambdaExceptionHelper;
-import java.util.ArrayList;
+import de.fraunhofer.iosb.ilt.faaast.service.util.StringHelper;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.stream.Collectors;
-import net.thisptr.jackson.jq.BuiltinFunctionLoader;
-import net.thisptr.jackson.jq.JsonQuery;
-import net.thisptr.jackson.jq.Scope;
-import net.thisptr.jackson.jq.Versions;
-import org.apache.commons.lang3.StringUtils;
 
 
 /**
@@ -71,62 +67,6 @@ public class JsonFormat implements Format {
     }
 
 
-    private String executeQuery(String value, String query) throws AssetConnectionException {
-        List<String> results = query.startsWith("jq|")
-                ? executeJqQuery(value, query.substring(3))
-                : executeJsonPathQuery(value, query);
-        if (results.isEmpty()) {
-            throw new AssetConnectionException(String.format("Query expression did not return any value (JSON path: %s, JSON: %s)", query, value));
-        }
-        if (results.size() > 1) {
-            throw new AssetConnectionException(String.format("Query expression returned more than one value (JSON path: %s, JSON: %s)", query, value));
-        }
-        return results.get(0);
-    }
-
-
-    private List<String> executeJqQuery(String value, String query) throws AssetConnectionException {
-        try {
-            ObjectMapper mapper = new ObjectMapper();
-            JsonNode inputNode = mapper.readTree(value);
-            Scope scope = Scope.newEmptyScope();
-            BuiltinFunctionLoader.getInstance().loadFunctions(Versions.JQ_1_7, scope);
-            JsonQuery q = JsonQuery.compile(query, Versions.JQ_1_7);
-
-            final List<JsonNode> results = new ArrayList<>();
-            q.apply(scope, inputNode, results::add);
-            return results.stream()
-                    .map(Object::toString)
-                    .toList();
-        }
-        catch (JsonProcessingException e) {
-            throw new AssetConnectionException(String.format("error executing JQ query (query: %s, JSON: %s)", query, value), e);
-        }
-    }
-
-
-    private List<String> executeJsonPathQuery(String value, String query) throws AssetConnectionException {
-        try {
-            List<Object> results = JsonPath
-                    .using(Configuration.defaultConfiguration().addOptions(Option.ALWAYS_RETURN_LIST))
-                    .parse(value)
-                    .read(query);
-            return results.stream()
-                    .map(Object::toString)
-                    .toList();
-        }
-        catch (PathNotFoundException e) {
-            throw new AssetConnectionException(String.format("value addressed by JSONPath not found (JSON path: %s, JSON: %s)", query, value), e);
-        }
-        catch (InvalidPathException e) {
-            throw new AssetConnectionException(String.format("invalid JSONPath (JSON path: %s)", query), e);
-        }
-        catch (JsonPathException e) {
-            throw new AssetConnectionException(String.format("error resolving JSONPath (JSON path: %s, JSON: %s)", query, value), e);
-        }
-    }
-
-
     @Override
     public Map<String, DataElementValue> read(String value, Map<String, ElementInfo> elements) throws AssetConnectionException {
         if (elements == null) {
@@ -140,8 +80,29 @@ public class JsonFormat implements Format {
                 LambdaExceptionHelper.rethrowFunction(x -> {
                     String query = x.getValue().getQuery();
                     String actualValue = value;
-                    if (!StringUtils.isBlank(query)) {
-                        actualValue = executeQuery(value, query);
+                    if (!StringHelper.isBlank(query)) {
+                        try {
+                            List<Object> jsonPathResult = JsonPath
+                                    .using(Configuration.defaultConfiguration().addOptions(Option.ALWAYS_RETURN_LIST))
+                                    .parse(value)
+                                    .read(query);
+                            if (jsonPathResult.isEmpty()) {
+                                throw new AssetConnectionException(String.format("JSONPath expression did not return any value (JSON path: %s, JSON: %s)", query, value));
+                            }
+                            if (jsonPathResult.size() > 1) {
+                                throw new AssetConnectionException(String.format("JSONPath expression returned more than one value (JSON path: %s, JSON: %s)", query, value));
+                            }
+                            actualValue = jsonPathResult.get(0).toString();
+                        }
+                        catch (PathNotFoundException e) {
+                            throw new AssetConnectionException(String.format("value addressed by JSONPath not found (JSON path: %s, JSON: %s)", query, value), e);
+                        }
+                        catch (InvalidPathException e) {
+                            throw new AssetConnectionException(String.format("invalid JSONPath (JSON path: %s)", query), e);
+                        }
+                        catch (JsonPathException e) {
+                            throw new AssetConnectionException(String.format("error resolving JSONPath (JSON path: %s, JSON: %s)", query, value), e);
+                        }
                     }
                     try {
                         TypeInfo<?> typeInfo = x.getValue().getTypeInfo();
@@ -162,10 +123,24 @@ public class JsonFormat implements Format {
     }
 
 
+    private static DataElementValue handleException(String message, Exception e, boolean shouldThrowException) throws AssetConnectionException {
+        if (shouldThrowException) {
+            throw new AssetConnectionException(message, e);
+        }
+        return null;
+    }
+
+
     @Override
     public String write(DataElementValue value) throws AssetConnectionException {
         try {
-            return serializer.write(value);
+            return serializer.write(
+                    value,
+                    new OutputModifier.Builder()
+                            .content(Content.VALUE)
+                            .level(Level.DEEP)
+                            .extend(Extent.WITH_BLOB_VALUE)
+                            .build());
         }
         catch (SerializationException | UnsupportedModifierException e) {
             throw new AssetConnectionException("serializing value to JSON failed", e);